refactor: move LessonMetadata, HeaderListing and LessonOverviewDetails
