+++ src/pages/teachers/programmes/[programmeSlug]/units.tsx
@@ -30,7 +30,7 @@ import { RESULTS_PER_PAGE } from "@/utils/resultsPerPage";
 import getPageProps from "@/node-lib/getPageProps";
 import curriculumApi2023 from "@/node-lib/curriculum-api-2023";
 import { filterLearningTheme } from "@/utils/filterLearningTheme/filterLearningTheme";
-import HeaderListing from "@/components/HeaderListing/HeaderListing";
+import HeaderListing from "@/components/TeacherComponents/HeaderListing/HeaderListing";
 import isSlugLegacy from "@/utils/slugModifiers/isSlugLegacy";
 import useAnalytics from "@/context/Analytics/useAnalytics";
 import useAnalyticsPageProps from "@/hooks/useAnalyticsPageProps";
